BUG FIX - remove error message from login_required method
Before if you tried ot access a page you did not have access to you'd be redirected to the homepage and the error message would show that you must be logged in. I removed this because we redirect from the home page to "login" when you visit the main url and it was triggering an error message every time someone visited the site. Better way is to only render the error message on some pages. Will look into this next.